requirements.txt
asgiref==3.2.7
Django==3.0.6
django-filter==2.2.0
djangorestframework==3.11.0
mysqlclient==1.4.6
protobuf==3.12.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
